Improve drop and anti-afk

+++ minecraft/neko/module/modules/misc/AutoCmd.java
@@ -5,7 +5,7 @@ import neko.module.Module;
 import neko.utils.ChatUtils;
 
 public class AutoCmd extends Module {
-	public static int sec = 5;
+	public static double sec = 5;
 	public static String cmd = "";
 	public int count = 0;
